requirements.txt
Django==4.1.7
djangorestframework==3.14.0
django-filter==23.1
djangorestframework-simplejwt==5.2.2

